[Fix#46]: Seperate_CSR_SSR
